chore: move all common sources + first packaging commit
Common sources are now in a separate crate. This has the advantage to avoid many "unused" warnings. Indeed as chsr uses some calls that sr don't use, it raise this warning. And as the same with sr. So many unused warnings were wrongly reported.

I started to make deployment work, as sr and chsr seems to be well-tested and made some security checks.

For now, deployment is WIP.