WIP deployment
This work aims to code the installation process to take into account many parameters that are complex to manipulate using shell, but also avoid some dependencies that are only used for the installation process.

This work should also simplify deployment on many distributions, while applying the principle of least privilege.

However, one last question is about the dependency management, as it requires rust, we still need a shell script to install rust toolchains. As capable requires nighly,  and sr requires stable>=1.70, it may requires both toolchains for a clean release binary.