chore: Remove CHANGELOG.md and update dependencies
This commit removes the CHANGELOG.md file and updates the dependencies in the Cargo.toml file. The CHANGELOG.md file was deleted as it was redundant with Github releases notes. The dependencies in the Cargo.toml file were updated to newer versions. This commit is a chore and does not introduce any functional changes.